activity 변수명 정리, class 정리, DB filename저장
