--- example/testApp1/app/src/main/java/com/example/testapp1/NetworkTask.java
@@ -1,27 +1,0 @@
-package com.example.testapp1;
-
-import android.content.ContentValues;
-import android.os.AsyncTask;
-
-public class NetworkTask extends AsyncTask<Void, Void, String> {
-
-    private String url;
-    private ContentValues values;
-
-    public NetworkTask(String url, ContentValues values) {
-
-        this.url = url;
-        this.values = values;
-    }
-
-    @Override
-    protected String doInBackground(Void... params) {
-
-        String result; // 요청 결과를 저장할 변수.
-        RequestHttpURLConnection requestHttpURLConnection = new RequestHttpURLConnection();
-        result = requestHttpURLConnection.request(url, values); // 해당 URL로 부터 결과물을 얻어온다.
-
-        return result;
-    }
-
-}
