+++ example/testApp1/app/src/main/res/layout/activity_main.xml
@@ -13,7 +13,7 @@
         android:text="이름 : " />
 
     <EditText
-        android:id="@+id/EditUserName"
+        android:id="@+id/activity_main_et_user_name"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:hint="입력 해 주세요"
@@ -32,7 +32,7 @@
         >
 
         <Button
-            android:id="@+id/BtnGenderMan"
+            android:id="@+id/activity_main_btn_gender_man"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:background="@color/white"
@@ -41,7 +41,7 @@
             android:textStyle="bold" />
 
         <Button
-            android:id="@+id/BtnGenderWoman"
+            android:id="@+id/activity_main_btn_gender_woman"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:background="@color/white"
@@ -58,7 +58,7 @@
         android:text="나이 : " />
 
     <Spinner
-        android:id="@+id/SpinnerAge"
+        android:id="@+id/activity_main_spn_age"
         android:layout_width="200sp"
         android:layout_height="wrap_content"
         android:textSize="18sp"
@@ -72,7 +72,7 @@
         android:text="핸드폰번호 : " />
 
     <EditText
-        android:id="@+id/EditPhoneNumber"
+        android:id="@+id/activity_main_et_phone_number"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:inputType="phone"
@@ -87,28 +87,21 @@
         >
 
     <Button
-        android:id="@+id/BtnSave"
+        android:id="@+id/activity_main_btn_save"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="저장"
+        android:text="다 음"
         />
 
         <Button
-            android:id="@+id/BtnUpload"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="파일 업로드"
-            />
-
-        <Button
-            android:id="@+id/BtnUploadList"
+            android:id="@+id/activity_main_btn_upload_list"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="파일 업로드 목록"
             />
 
         <Button
-            android:id="@+id/BtnList"
+            android:id="@+id/activity_main_btn_list"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="회원 목록"
