+++ example/testApp1/app/src/main/res/layout/file_upload_activity.xml
@@ -11,21 +11,27 @@
         android:layout_height="wrap_content">
 
     <Button
-        android:id="@+id/bt_choose"
-        android:text="Choose"
+        android:id="@+id/file_upload_activity_btn_back"
+        android:text="뒤로 가기"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content" />
 
     <Button
-        android:id="@+id/bt_upload"
-        android:text="Upload"
+        android:id="@+id/file_upload_activity_btn_choose"
+        android:text="사진 선택"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content" />
+
+    <Button
+        android:id="@+id/file_upload_activity_btn_regist"
+        android:text="등 록"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content" />
 
     </LinearLayout>
 
     <ImageView
-        android:id="@+id/iv_preview"
+        android:id="@+id/file_upload_activity_iv_preview"
         android:layout_width="match_parent"
         android:layout_height="wrap_content" />
